% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{func}
\alias{func}
\title{Surprise Function}
\usage{
func(x)
}
\arguments{
\item{x}{integer, any integer, usually a number from 1 to 10}
}
\value{
returns a message associated to the number provided in x
}
\description{
Surprise Function
}
