% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_load.R
\name{func_load}
\alias{func_load}
\title{Load default parameter values}
\usage{
func_load(fun, envir = globalenv())
}
\arguments{
\item{fun}{Function object, not character name of function}

\item{envir}{Environment to load default parameters in to.
Default: `globalenv()`}
}
\description{
Of a function or list of functions
}
\examples{
\dontrun{
# A single function
foo <- function(x=5, y=4) x+y
func_load(foo)
x; y

# Many functions
foo <- function(x=5, y=4) x+y
bar <- function(a="a", b="b") c(a,b)
func_load(c(foo, bar))
x; y; a; b
}
}
