% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_load.R
\name{func_load}
\alias{func_load}
\title{Load default parameter values}
\usage{
func_load(fun, envir = globalenv())
}
\arguments{
\item{fun}{Function object, not character name of function}

\item{envir}{Environment to load default parameters in to.
Default: `globalenv()`}
}
\description{
For a function or list of functions
}
\examples{
\dontrun{
# one function
foo <- function(w=5, y=4) w+y
func_load(foo)
w; y

# many functions
foo <- function(w=5, y=4) w+y
bar <- function(a="a", b="b") c(a,b)
func_load(c(foo, bar))
w; y; a; b
}
}
