% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_load_path.R
\name{func_load_path}
\alias{func_load_path}
\title{Load default parameter values}
\usage{
func_load_path(path, envir = globalenv())
}
\arguments{
\item{path}{Path to a file with one or more functions.
required}

\item{envir}{Environment to load default parameters in to.
Default: .GlobalEnv}
}
\description{
From a file
}
\examples{
\dontrun{
tmpf <- tempfile(fileext=".R")
cat("fun <- function(bar = 5) bar + 1\n", file=tmpf)
func_load_path(path=tmpf)
bar
}
}
